--- src/ButtonNavigation.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import { withStyles } from 'material-ui/styles';
-import BottomNavigation, { BottomNavigationAction } from 'material-ui/BottomNavigation';
-
-const styles = {
-  root: {
-    width: 300,
-  },
-};
-
-class SimpleBottomNavigation extends React.Component {
-  state = {
-    value: 0,
-  };
-
-  handleChange = (event, value) => {
-    this.setState({ value });
-  };
-
-  render() {
-    const { classes } = this.props;
-    const { value } = this.state;
-
-    return (
-      <BottomNavigation
-        value={value}
-        onChange={this.handleChange}
-        showLabels
-        className={classes.root}
-      >
-        <BottomNavigationAction label="My Chats"/>
-        <BottomNavigationAction label="Explore"/>
-      </BottomNavigation>
-    );
-  }
-}
-
-export default withStyles(styles)(SimpleBottomNavigation);
